refactor: 인수 테스트 패키지 변경
acceptance 패키지 한 곳에서 쉽게 관리하기위해 추가했다.
