+++ src/test/java/subway/acceptance/AbstractAcceptanceTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ValidatableResponse;
